Backend/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
sqlalchemy>=2.0.25
pydantic>=2.5.3
pydantic-settings>=2.1.0
python-multipart>=0.0.6
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt>=4.1.2
psycopg2-binary>=2.9.9
requests>=2.31.0
pandas>=2.2.0
openpyxl==3.1.2
email-validator==2.1.0
openai>=1.0.0
pinecone-client>=3.0.0
google-generativeai>=0.3.0
pypdf
anthropic>=0.18.0
gunicorn
reportlab
python-docx
